remove duplicate Activity and ActivitySubTypes
Activity.rb and FFS_Activity.rb files both contained Activity and
ActivitySubTypes.  The ones in FFS_Activity were older and didn't
contain all of the entries that are in Activity.rb. Let's just remove
what's in FFS_Activity.rb and reference what's in Activity.rb. This
issue fixes the following entry on the generated HTML page:

    Sub Type: Undefined gravel_cycling

With this change, it now correctly says:

    Sub Type: Gravel Cycling